The module for each class created
